Update lib/template/Sortable.php line 108
The function 'getFinalPosition()' needs to be called on the Doctrine_Template_Sortable object since this class has the magical function. Else you would get an error saying the function does not exist on the $object which would require implementation of the method and that's not desirable since it's a behavior and should work out of the box.